Split off regex matching, add a test for it
